approval-worker/pyproject.toml
[project]
name = "sentinel-health-approval-worker"
version = "0.1.0"
description = "Sentinel-Health Approval Worker — Pub/Sub triggered HITL approval service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-cloud-firestore>=2.19.0",
    "google-cloud-pubsub>=2.27.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
